+++ src/pages/search/components/SearchBar/index.css.ts
